move logic for pick up place, to the PlacePickerItem
