+++ apps/core/scenes/search/PlacePicker/PlacePickerList.js
@@ -5,13 +5,11 @@ import { ScrollView } from 'react-native';
 import { StyleSheet } from '@kiwicom/universal-components';
 import { graphql, createFragmentContainer } from '@kiwicom/margarita-relay';
 
-import { type Location } from '../SearchContext';
 import PlaceItem from './PlaceItem';
 import type { PlacePickerList_locations as PlacePickerListTypes } from './__generated__/PlacePickerList_locations.graphql';
 
 type Props = {|
   +locations: ?PlacePickerListTypes,
-  +onPressItem: Location => void,
 |};
 
 const PlacePickerList = (props: Props) => {
@@ -19,11 +17,7 @@ const PlacePickerList = (props: Props) => {
   return (
     <ScrollView style={styles.container}>
       {locations.map(location => (
-        <PlaceItem
-          key={location?.node?.id}
-          item={location?.node}
-          onPress={props.onPressItem}
-        />
+        <PlaceItem key={location?.node?.id} item={location?.node} />
       ))}
     </ScrollView>
   );
